Enum ketos::compile::CompileError

pub enum CompileError {
    ArityError {
        name: Name,
        expected: Arity,
        found: u32,
    },
    CannotDefine(Name),
    ConstantExists(Name),
    DuplicateExports,
    DuplicateModuleDoc,
    DuplicateParameter(Name),
    ExportError {
        module: Name,
        name: Name,
    },
    ImportCycle(Name),
    ImportError {
        module: Name,
        name: Name,
    },
    ImportShadow(Name),
    InvalidCallExpression(&'static str),
    InvalidCommaAt,
    InvalidModuleName(Name),
    MacroRecursionExceeded,
    MissingExport,
    ModuleError(Name),
    NotConstant(Name),
    OperandOverflow(u32),
    PrivacyError {
        module: Name,
        name: Name,
    },
    SyntaxError(&'static str),
    UnbalancedComma,
}

Represents an error generated while compiling to bytecode.

Variants
ArityError

Error in arity for call to system function

Fields of ArityError

name: Name

Name of function

expected: Arity

Expected count or range of arguments

found: u32

Number of arguments present

CannotDefine(Name)

Attempt to define name of standard value or operator

ConstantExists(Name)

Attempt to define name held by const value

DuplicateExports

Duplicate exports declaration

DuplicateModuleDoc

Duplicate module-doc declaration

DuplicateParameter(Name)

Duplicate name in parameter list

ExportError

Attempt to export nonexistent name from module

Fields of ExportError

module: Name

Module name

name: Name

Imported name

ImportCycle(Name)

Recursion in module imports

ImportError

Attempt to import nonexistent name from module

Fields of ImportError

module: Name

Module name

name: Name

Imported name

ImportShadow(Name)

Attempt to import or define name which already exists

InvalidCallExpression(&'static str)

Invalid expression to function call

InvalidCommaAt

,@expr form outside of a list

InvalidModuleName(Name)

Module name contains invalid characters

MacroRecursionExceeded

Recursion limit exceeded while expanding macros

MissingExport

Missing export declaration in loaded module

ModuleError(Name)

Failed to load a module

NotConstant(Name)

const operator value is not constant

OperandOverflow(u32)

Operand value overflow

PrivacyError

Attempt to import value that is not exported

Fields of PrivacyError

module: Name

Module name

name: Name

Imported name

SyntaxError(&'static str)

Error in parsing operator syntax

UnbalancedComma

More commas than backquotes