Enum ketos::encode::EncodeError

pub enum EncodeError {
    InvalidUtf8,
    Overflow,
    UnencodableType(&'static str),
    UnencodableValue(&'static str),
}

Error in encoding bytecode file format

Variants
InvalidUtf8

A Path value contains invalid UTF-8

Overflow

Integer overflow in encoding value

UnencodableType(&'static str)

Attempt to encode a type that cannot be encoded

UnencodableValue(&'static str)

Attempt to encode a value that cannot be encoded