Enum ketos::error::Error

pub enum Error {
    CompileError(CompileError),
    DecodeError(DecodeError),
    EncodeError(EncodeError),
    ExecError(ExecError),
    IoError(IoError),
    ParseError(ParseError),
    RestrictError(RestrictError),
    Custom(Box<StdError>),
}

Consolidated error type; contains one of a category of errors.

Variants
CompileError(CompileError)

Error in compiling code to bytecode

DecodeError(DecodeError)

Error in decoding bytecode file format

EncodeError(EncodeError)

Error in encoding bytecode file format

ExecError(ExecError)

Error in executing code

IoError(IoError)

Error in file I/O operation

ParseError(ParseError)

Error in scanning text or parsing syntax

RestrictError(RestrictError)

Code execution breached configured restrictions

Custom(Box<StdError>)

Customized error value implementing std::error::Error

Methods

impl Error

fn custom<E: 'static + StdError>(e: E) -> Error

Returns an Error value wrapping a custom error type.

fn description(&self) -> &'static str

Returns a string describing the nature of the error.