Module ketos::function

Contains implementations of core system functions.

Structs

Function
Represents a function implemented in Rust.

Lambda
Represents a function that evaluates an expression.

SystemFn
Represents a system function

Enums

Arity
Describes the number of arguments a function may accept.

Statics

SYSTEM_FNS
System function implementations.

Functions

add_number
Returns the result of adding two values together.

div_number
Returns the result of dividing two values.

first
Returns the first element of a list or string.

floor_div_number_step
Returns the result of floor-dividing two values, without calling floor on the result.

floor_number
Returns a value rounded toward negative infinity.

init
Returns all but the last element of a list or string.

last
Returns the last element of a list or string.

mul_number
Returns the result of multiplying two values together.

neg_number
Returns the result of negating a value.

plural
Returns the suitable plural suffix "" or "s" for count n.

sub_number
Returns the resulting of subtracting a value from another.

tail
Returns all but the first element of a list or string.

Type Definitions

FunctionImpl
SystemFn implemented by Rust function