Module ketos::integer

Arbitrary precision integer and ratio types.

Structs

FromStrIntError
Error produced when failing to parse an Integer from &str.

FromStrRadixError
Error produced when failing to parse an Integer with radix from &str.

FromStrRatioError
Error produced when failing to parse a Ratio from &str.

Integer
Arbitrary precision signed integer

Ratio
Arbitrary precision signed integer ratio

Enums

Sign
A Sign is a BigInt's composing element.