Struct ketos::integer::Integer

pub struct Integer(_);

Arbitrary precision signed integer

Methods

impl Integer

fn from_bytes_be(sign: Sign, bytes: &[u8]) -> Integer

Creates an Integer from a sign and a series of big-endian bytes.

fn from_bytes_le(sign: Sign, bytes: &[u8]) -> Integer

Creates an Integer from a sign and a series of little-endian bytes.

fn from_f64(f: f64) -> Option<Integer>

Creates an Integer with the value of the given f64. Returns None if the value cannot be converted.

fn from_i8(i: i8) -> Integer

Creates an Integer with the value of the given i8.

fn from_i16(i: i16) -> Integer

Creates an Integer with the value of the given i16.

fn from_i32(i: i32) -> Integer

Creates an Integer with the value of the given i32.

fn from_i64(i: i64) -> Integer

Creates an Integer with the value of the given i64.

fn from_isize(i: isize) -> Integer

Creates an Integer with the value of the given isize.

fn from_u8(i: u8) -> Integer

Creates an Integer with the value of the given u8.

fn from_u16(i: u16) -> Integer

Creates an Integer with the value of the given u16.

fn from_u32(i: u32) -> Integer

Creates an Integer with the value of the given u32.

fn from_u64(i: u64) -> Integer

Creates an Integer with the value of the given u64.

fn from_usize(u: usize) -> Integer

Creates an Integer with the value of the given usize.

fn from_str_radix(s: &str, radix: u32) -> Result<Integer, FromStrRadixError>

Returns an Integer represented by a string in the given radix. radix must be in the range [2, 36].

fn to_bytes_be(&self) -> (Sign, Vec<u8>)

Returns integer sign and a series of big-endian bytes.

fn to_bytes_le(&self) -> (Sign, Vec<u8>)

Returns integer sign and a series of little-endian bytes.

fn to_str_radix(&self, radix: u32) -> String

Returns a string representation of the Integer in the given radix. radix must be in the range [2, 36].

fn to_i8(&self) -> Option<i8>

Returns the Integer as an i8 value.

fn to_i16(&self) -> Option<i16>

Returns the Integer as an i16 value.

fn to_i32(&self) -> Option<i32>

Returns the Integer as an i32 value.

fn to_i64(&self) -> Option<i64>

Returns the Integer as an i64 value.

fn to_isize(&self) -> Option<isize>

Returns the Integer as an isize value.

fn to_u8(&self) -> Option<u8>

Returns the Integer as an u8 value.

fn to_u16(&self) -> Option<u16>

Returns the Integer as an u16 value.

fn to_u32(&self) -> Option<u32>

Returns the Integer as an u32 value.

fn to_u64(&self) -> Option<u64>

Returns the Integer as an u64 value.

fn to_usize(&self) -> Option<usize>

Returns the Integer as an usize value.

fn to_f32(&self) -> Option<f32>

Returns the Integer as an f32 value.

fn to_f64(&self) -> Option<f64>

Returns the Integer as an f64 value.

fn pow(self, exp: usize) -> Integer

Raises the value to the power of exp.

fn abs(&self) -> Integer

Returns the absolute value of an Integer.

fn is_multiple_of(&self, rhs: &Integer) -> bool

Returns whether self is a multiple of rhs.

fn is_negative(&self) -> bool

Returns whether the Integer is less than zero.

fn is_positive(&self) -> bool

Returns whether the Integer is greater than zero.

fn is_zero(&self) -> bool

Returns whether the Integer is equal to zero.

fn zero() -> Integer

Returns an Integer of the value zero.

fn is_one(&self) -> bool

Returns whether the Integer is equal to one.

fn one() -> Integer

Returns an Integer of the value one.