Module ketos::module

Implements loading named values from code modules.

Structs

BuiltinModuleLoader
Loads builtin modules.

ChainModuleLoader
A chained module loader contains two ModuleLoader implementations.

FileModuleLoader
Loads modules from source files and compiled bytecode files.

Module
Contains the values in a loaded module's namespace.

ModuleBuilder
Helper to build modules in Rust code.

ModuleCode
Contains code and data representing a compiled module

ModuleRegistry
Loads modules into the running program and caches previously loaded modules

NullModuleLoader
Loads no modules.

Constants

COMPILED_FILE_EXTENSION
File extension for ketos compiled bytecode files.

FILE_EXTENSION
File extension for ketos source files.

Traits

ModuleLoader
Provides a method for loading modules into a running interpreter.