Struct ketos::module::FileModuleLoader

pub struct FileModuleLoader { /* fields omitted */ }

Loads modules from source files and compiled bytecode files.

Methods

impl FileModuleLoader

fn new() -> FileModuleLoader

Creates a new FileModuleLoader that will search the current directory for modules.

fn with_search_paths(paths: Vec<PathBuf>) -> FileModuleLoader

Creates a new FileModuleLoader that will search the given series of directories to load modules.

fn add_search_path(&mut self, path: PathBuf)

Adds a directory to search for module files.

fn set_read_bytecode(&mut self, set: bool)

Sets whether the FileModuleLoader will search for compiled bytecode files when loading modules. The default is true.

fn set_write_bytecode(&mut self, set: bool)

Sets whether the FileModuleLoader will write compiled bytecode files after loading a module from source. The default is true.