Struct ketos::name::NameMapSlice

pub struct NameMapSlice<T> {
    // some fields omitted
}

Maps names to values in a sorted boxed slice.

Values may overwrite existing values, but new names cannot be inserted.

Methods

impl<T> NameMapSlice<T>

fn contains_key(&self, name: Name) -> bool

Returns whether the map contains a value for the given name.

fn get(&self, name: Name) -> Option<&T>

Returns the value corresponding to the given name.

fn get_values(&self) -> &[(Name, T)]

Returns a slice of the contained names and values.

fn set(&mut self, name: Name, value: T) -> Option<T>

Overwrites the value for the given name.

Returns None if the name does not exist in the mapping. When None is returned, no value will have been stored in the mapping.

fn is_empty(&self) -> bool

Returns whether the given map is empty.

fn iter(&self) -> Iter<(Name, T)>

Returns an iterator over names and values.

fn into_name_map(self) -> NameMap<T>

Elevates the map into NameMap, which may receive new key-value pairs.

fn len(&self) -> usize

Returns the number of name-value pairs contained in the map.