Struct ketos::name::NameStore

pub struct NameStore {
    // some fields omitted
}

Maps interned Name values to their String representations

Methods

impl NameStore

fn new() -> NameStore

Constructs an empty NameStore.

fn add(&mut self, name: &str) -> Name

Adds a name to the NameStore if it is not present. Returns a Name value to refer to the new or existing name.

fn get_name(&self, name: &str) -> Option<Name>

Returns the Name value of a given string, if it exists.

fn get(&self, name: Name) -> &str

Returns the string representation of an interned name.

fn iter(&self) -> Iter<String>

Iterates over all stored names.