Function ketos::pretty::pretty_print

pub fn pretty_print<W: Write>(w: &mut W, names: &NameStore, v: &Value, indent: u32) -> Result

Writes a human-readable representation of a Value to the given fmt::Write.

indent specifies the base indentation for items contained within lists. Indentation is not applied to the top level value.