Struct ketos::rc_vec::RcVec

pub struct RcVec<T> { /* fields omitted */ }

Represents a reference-counted view into a Vec. Subslices may be created which will share the underlying data buffer.

Methods

impl<T> RcVec<T>

fn new(data: Vec<T>) -> RcVec<T>

Constructs a new RcVec from a Vec.

fn is_empty(&self) -> bool

Returns whether the RcVec is empty.

fn len(&self) -> usize

Returns the number of elements visible to the RcVec.

fn slice<R: RangeArgument<usize>>(&self, range: R) -> RcVec<T>

Returns a subslice of the RcVec, with the range being relative to this slice's boundaries.

impl<T: Clone> RcVec<T>

fn into_vec(self) -> Vec<T>

Consumes the RcVec and returns the contained Vec. This will clone the contained values unless the data was uniquely held.

fn push(&mut self, t: T)

Pushes a value into the contained Vec.