Enum ketos::scope::MasterScope

pub enum MasterScope {}

Represents the universal namespace containing built-in symbols which are available in any context.

Methods

impl MasterScope

fn contains(name: Name) -> bool

Returns whether the given name corresponds to a value in master scope.

fn can_define(name: Name) -> bool

Returns whether the given name can be defined in global scope.

fn get(name: Name) -> Option<Value>

Returns a value corresponding to the given name in master scope.

fn names() -> MasterNames

Returns an iterator over all standard names.

fn values() -> MasterValues

Returns an iterator over all names contained in master scope.