Module ketos::trace

Provides facilities for expressing and storing tracebacks.

When a compilation or execution operation returns an error, it will store a traceback within thread-local storage. To access this value, the functions get_traceback (which clones the value) and take_traceback (which removes the value) can be used.

Structs

Trace
Represents a series of items, beginning with the outermost context and culminating with the context in which an error was generated.

Enums

TraceItem
Represents a single traceable event in either compilation or execution of code.

Functions

clear_traceback
Removes the traceback value for the current thread.

get_traceback
Clones and returns the traceback value for the current thread.

set_traceback
Assigns a traceback value for the current thread.

take_traceback
Removes and returns the traceback value for the current thread.