Enum ketos::trace::TraceItem

pub enum TraceItem {
    CallCode(Name, Name),
    CallExpr(Name),
    CallLambda(Name),
    CallMacro(Name, Name),
    CallOperator(Name, Name),
    CallSys(Name),
    Define(Name, Name),
    DefineConst(Name, Name),
    DefineLambda(Name),
    DefineMacro(Name, Name),
    DefineStruct(Name, Name),
    UseModule(Name, Name),
}

Represents a single traceable event in either compilation or execution of code.

Variants
CallCode(Name, Name)

Call to a code object; (scope name, code name)

CallExpr(Name)

Call to a code object generated by an expression

CallLambda(Name)

Call to an anonymous function

CallMacro(Name, Name)

Call to a macro; (scope name, macro name)

CallOperator(Name, Name)

Expansion of an operator; (scope name, operator name)

CallSys(Name)

Call to a system function

Define(Name, Name)

Definition of a named value; (scope name, definition name)

DefineConst(Name, Name)

Definition of a constant value; (scope name, const name)

DefineLambda(Name)

Definition of an anonymous lambda

DefineMacro(Name, Name)

Definition of a macro; (scope name, macro name)

DefineStruct(Name, Name)

Definition of a structure; (scope name, struct name)

UseModule(Name, Name)

Module import declaration; (scope name, module name)