Struct ketos::trace::Trace

pub struct Trace { /* fields omitted */ }

Represents a series of items, beginning with the outermost context and culminating with the context in which an error was generated.

Methods

impl Trace

fn new(items: Vec<TraceItem>, expr: Option<Value>) -> Trace

Creates a new Trace from a series of items.

fn single(item: TraceItem, expr: Option<Value>) -> Trace

Creates a new Trace from a single item.

fn items(&self) -> &[TraceItem]

Returns the series of traced items.

fn expr(&self) -> Option<&Value>

Returns a borrowed reference to the optional contained expression.

fn take_expr(&mut self) -> Option<Value>

Takes the optional contained expression and returns it.