Module ketos::value

Represents any possible value type.

Structs

ForeignFn
Represents a foreign value that contains a callable function or closure

Struct
Represents a structure value containing named fields

StructDef
Represents the definition of a class of struct value

Enums

Value
Represents a value.

Traits

AnyValue
A helper trait that is necessary as long as Any::get_type_id is unstable.

ForeignValue
Represents a type of value defined outside the core interpreter.

FromValue
Consumes a Value and returns a Rust value

FromValueRef
Borrows a Rust value from a Value