Enum ketos::value::Value

pub enum Value {
    Unit,
    Unbound,
    Bool(bool),
    Float(f64),
    Integer(Integer),
    Ratio(Ratio),
    Struct(Rc<Struct>),
    StructDef(Rc<StructDef>),
    Name(Name),
    Keyword(Name),
    Char(char),
    String(RcString),
    Bytes(Bytes),
    Path(PathBuf),
    Quasiquote(Box<Value>, u32),
    Comma(Box<Value>, u32),
    CommaAt(Box<Value>, u32),
    Quote(Box<Value>, u32),
    List(RcVec<Value>),
    Function(Function),
    Lambda(Lambda),
    Foreign(Rc<ForeignValue>),
}

Represents a value.

Variants
Unit

()

Unbound

Placeholder for missing optional values; should not be used to represent any real value.

Bool(bool)

Boolean -- true or false

Float(f64)

Floating point number

Integer(Integer)

Signed arbitrary precision integer

Ratio(Ratio)

Arbitrary precision ratio

Struct(Rc<Struct>)

Struct value

StructDef(Rc<StructDef>)

Struct definition

Name(Name)

Literal name

Keyword(Name)

Keyword

Char(char)

Character

String(RcString)

String

Bytes(Bytes)

Byte string

Path(PathBuf)

Path

Quasiquote(Box<Value>, u32)

Quasiquoted value; quote depth MUST NEVER be zero.

Comma(Box<Value>, u32)

Comma'd value; comma depth MUST NEVER be zero.

CommaAt(Box<Value>, u32)

Comma-at'd value; quote depth MUST NEVER be zero.

Quote(Box<Value>, u32)

Quoted value; quote depth MUST NEVER be zero.

List(RcVec<Value>)

Series of one or more values. MUST NEVER be of length zero. Use Unit to represent empty lists.

Function(Function)

Function implemented in Rust

Lambda(Lambda)

Compiled bytecode function

Foreign(Rc<ForeignValue>)

Boxed value of a foreign type

Methods

impl Value

fn new_foreign<T: ForeignValue>(t: T) -> Value

Returns a value of a foreign type.

fn new_foreign_fn<F>(name: Name, f: F) -> Value where F: Any + Fn(&Context, &mut [Value]) -> Result<Value, Error>

Returns a value containing a foreign function.

fn compare(&self, rhs: &Value) -> Result<Ordering, ExecError>

Compares two values; returns an error if the values cannot be compared.

fn is_equal(&self, rhs: &Value) -> Result<bool, ExecError>

Tests two values for equality; returns an error if the values cannot be compared.

fn is_identical(&self, rhs: &Value) -> bool

Returns whether this value is identical to another. The notable difference between this and eq is that float NaN values will compare equal.

fn take(&mut self) -> Value

Replaces the value with Unit and returns the old value.

fn quasiquote(self, n: u32) -> Value

Returns the value, quasi-quoted.

Panics

If n would overflow.

fn comma(self, n: u32) -> Value

Returns the value, comma'd.

Panics

If n would overflow.

fn comma_at(self, n: u32) -> Value

Returns the value, comma-at'd.

Panics

If n would overflow.

fn quote(self, n: u32) -> Value

Returns the value, quoted.

Panics

If n would overflow.

fn type_name(&self) -> &'static str

Returns a string describing the type of the value.