Enum ketos::bytecode::Instruction

pub enum Instruction {
    Load(u32),
    LoadC(u32),
    UnboundToUnit(u32),
    GetDef(u32),
    Push,
    Unit,
    True,
    False,
    Const(u32),
    Store(u32),
    LoadPush(u32),
    LoadCPush(u32),
    GetDefPush(u32),
    UnitPush,
    TruePush,
    FalsePush,
    ConstPush(u32),
    SetDef(u32),
    List(u32),
    Quote(u32),
    Quasiquote(u32),
    Comma(u32),
    CommaAt(u32),
    BuildClosure(u32, u32),
    Jump(u32),
    JumpIf(u32),
    JumpIfBound(u32, u32),
    JumpIfNot(u32),
    JumpIfNull(u32),
    JumpIfNotNull(u32),
    JumpIfEq(u32),
    JumpIfNotEq(u32),
    JumpIfEqConst(u32, u32),
    JumpIfNotEqConst(u32, u32),
    Null,
    NotNull,
    Eq,
    NotEq,
    EqConst(u32),
    NotEqConst(u32),
    Not,
    Inc,
    Dec,
    Append,
    First,
    Tail,
    Init,
    Last,
    FirstPush,
    TailPush,
    InitPush,
    LastPush,
    CallSys(u32),
    CallSysArgs(u32, u32),
    CallConst(u32, u32),
    Call(u32),
    Apply(u32),
    CallSelf(u32),
    TailCall(u32),
    Skip(u32),
    Return,
}

Represents an instruction and any immediate parameters.

Any addition, deletion, or modification to this enum constitutes a breaking change to the bytecode format.

Variants

Load

Load a value from the stack

LoadC

Load a value from enclosed values

UnboundToUnit

If value n on the stack is Unbound, replace it with ().

GetDef

Load a named value from global scope

Push

Push a value onto the stack; value is invalidated.

Unit

Load () into value

True

Load true into value

False

Load false into value

Const

Load a const into value

Store

Store a value in the stack; value is invalidated.

LoadPush

Load, then push

LoadCPush

Load enclosed value, then push

GetDefPush

Load a name from global scope, then push

UnitPush

Load (), then push

TruePush

Load true, then push

FalsePush

Load false, then push

ConstPush

Load a const, then push

SetDef

Assign a value to a name into global scope

List

Build a list of n values from the stack

Quote

Transform value into an n-quoted value

Quasiquote

Transform value into an n-quasiquoted value

Comma

Transform value into an n-comma'd value

CommaAt

Transform value into an n-comma-at'd value

BuildClosure

Create a closure from code object in n_const and a list of n_values values on the stack; parameters are (n_const, n_values).

Jump

Jump to a label

JumpIf

Jump if value is true

JumpIfBound

Jump if value n from the stack is bound; parameters are (label, n)

JumpIfNot

Jump if value is false

JumpIfNull

Jump if value is ()

JumpIfNotNull

Jump if value is not ()

JumpIfEq

Jump if value is equal to top of stack

JumpIfNotEq

Jump if value is not equal to top of stack

JumpIfEqConst

Jump if value is equal to const; parameters are (label, n)

JumpIfNotEqConst

Jump if value is not equal to const; parameters are (label, n)

Null

Test whether value is ()

NotNull

Test whether value is not ()

Eq

Pops from the top of the stack and tests for equality with value.

NotEq

Pops from the top of the stack and tesst for inequality with value.

EqConst

Tests value for equality with const n

NotEqConst

Tests value for inequality with const n

Not

Negate boolean value

Inc

Increment integer value

Dec

Decrement integer value

Append

Append value to list on top of stack; result will be in value

First

Take first element of list and assign to value

Tail

Take tail of list and assign to value

Init

Take head of list and assign to value

Last

Take last element of list and assign to value

FirstPush

Push first value from list

TailPush

Push tail from list

InitPush

Push head from list

LastPush

Push last value from list

CallSys

Call system function n with known number of arguments on stack. Only functions with Exact arity may be called in this manner.

CallSysArgs

Call system function with n arguments on the stack; parameters are (sys_fn, n_args).

CallConst

Call const function with arguments on the stack; parameters are (const, n_args).

Call

Call function on the stack with n arguments from the top of the stack

Apply

Call function on the stack with n stack arguments and additional arguments from list value

CallSelf

Call current code object with n arguments from the top of the stack; this does not perform a tail call

TailCall

Perform tail-recursive call with n arguments from the top of the stack

Skip

Remove n values from the top of the stack

Return

Return value from function

Methods

impl Instruction

fn decode(r: &mut CodeReader) -> Result<Instruction, ExecError>

Decodes a single Instruction from a CodeReader.

fn encode(&self, w: &mut CodeBlock, short: bool) -> Result<(), CompileError>

Encodes a single Instruction into a CodeBlock.

fn get_jump_label(&self) -> Option<u32>

If the instruction is a jump instruction, returns the jump offset. Otherwise, returns None.

fn max_len() -> usize

Returns the maximum length, in bytes, of an encoded instruction.