Struct ketos::bytecode::Code

pub struct Code {
    pub name: Option<Name>,
    pub consts: Box<[Value]>,
    pub code: Box<[u8]>,
    pub kw_params: Box<[Name]>,
    pub n_params: u32,
    pub req_params: u32,
    pub flags: u32,
}

Represents a compiled bytecode function or expression.

Fields

name

Code object name, if present; top-level expressions and lambda code values do not have a name.

consts

Const values referenced in bytecode

code

Function body bytecode

kw_params

Names of keyword parameters accepted in the order in which they are expected.

n_params

Number of positional parameters accepted; this includes optional parameters, but excludes keyword and rest parameters, if accepted. Optional parameter values may be Unbound.

req_params

Number of positional parameters which must not be Unbound.

flags

Miscellaneous flags; see code_flags for bit flag values.

Methods

impl Code

fn arity(&self) -> Arity

Returns the computed arity of the compiled function.

fn has_rest_params(&self) -> bool

Returns whether the function accepts a rest parameter.

fn has_kw_params(&self) -> bool

Returns whether the function accepts one or more keyword parameters.