Enum ketos::exec::ExecError

pub enum ExecError {
    ArityError {
        name: Option<Name>,
        expected: Arity,
        found: u32,
    },
    CompareNaN,
    CannotCompare(&'static str),
    CannotDefine(Name),
    DivideByZero,
    DuplicateField(Name),
    DuplicateKeyword(Name),
    DuplicateStructDef(Name),
    FieldError {
        struct_name: Name,
        field: Name,
    },
    FieldTypeError {
        struct_name: Name,
        field: Name,
        expected: Name,
        found: &'static str,
    },
    FormatError {
        fmt: Box<str>,
        span: Span,
        err: FormatError,
    },
    InvalidClosureValue(u32),
    InvalidConst(u32),
    InvalidDepth,
    InvalidJump(u32),
    InvalidSlice(usize, usize),
    InvalidStack(u32),
    InvalidSystemFn(u32),
    MissingArgCount(Name),
    MissingField {
        struct_name: Name,
        field: Name,
    },
    NameError(Name),
    NotCharBoundary(usize),
    OddKeywordParams,
    OutOfBounds(usize),
    Overflow,
    Panic(Option<Value>),
    StackOverflow,
    StructDefError(Name),
    TypeError {
        expected: &'static str,
        found: &'static str,
    },
    StructMismatch {
        lhs: Name,
        rhs: Name,
    },
    TypeMismatch {
        lhs: &'static str,
        rhs: &'static str,
    },
    UnexpectedEnd,
    UnrecognizedKeyword(Name),
    UnrecognizedOpCode(u8),
}

Represents an error generated while executing bytecode.

Variants

ArityError

Error in arity to function call

Fields

name

Name of function, if available

expected

Expected count or range of arguments

found

Number of arguments present

CompareNaN

Attempt to compare with a NaN Float value.

CannotCompare

Type does not support ordered comparison

CannotDefine

Attempt to redefine a name in master scope

DivideByZero

Attempt to divide by a number equal to zero.

DuplicateField

Duplicate field name in struct definition

DuplicateKeyword

Duplicate keyword argument to function

DuplicateStructDef

Duplicate struct definition

FieldError

No such field name in struct

Fields

struct_name

Name of struct type

field

Field name

FieldTypeError

Type error assigning value to field

Fields

struct_name

Name of struct type

field

Name of field

expected

Expected type name

found

Name of type received

FormatError

Error in format call

Fields

fmt

Supplied format string

span

Span within format string

err

Formatting error produced

InvalidClosureValue

Invalid index into closure values

InvalidConst

Invalid const index

InvalidDepth

Invalid (zero) depth value to Quote, Quasiquote, or Comma instruction

InvalidJump

Invalid jump label

InvalidSlice

Slice indices out of order

InvalidStack

Invalid stack index

InvalidSystemFn

Invalid system function

MissingArgCount

CallSys instruction for system function which requires argument count

MissingField

Attempt to construct a Struct without the given field

Fields

struct_name

Struct type name

field

Field name

NameError

Attempt to lookup a name that did not exist in scope.

NotCharBoundary

Attempt to slice a string not along UTF-8 code point boundaries.

OddKeywordParams

Odd number of parameters when keyword-value pairs expected

OutOfBounds

Attempt to access an element in a list that is out of bounds.

Overflow

Integer overflow during certain arithmetic operations.

Panic

Code called panic

StackOverflow

Exceeded maximum stack size

StructDefError

Struct definition not found

TypeError

Operation performed on unexpected type

Fields

expected

Name of the type expected

found

Name of the type received

StructMismatch

Function received a value of incorrect type

Fields

lhs

Type of left-hand side value

rhs

Type of right-hand side value

TypeMismatch

Attempt to operate on two values of incompatible types

Fields

lhs

Type of left-hand side value

rhs

Type of right-hand side value

UnexpectedEnd

Unexpected end in bytecode

UnrecognizedKeyword

Unrecognized keyword passed to function

UnrecognizedOpCode

Unrecognized opcode

Methods

impl ExecError

fn expected(expected: &'static str, v: &Value) -> ExecError

Convenience function to return a TypeError value when expected type is expected, but some other type of value is found.