Struct ketos::integer::Ratio

pub struct Ratio(_);

Arbitrary precision signed integer ratio

Methods

impl Ratio

fn new(numer: Integer, denom: Integer) -> Ratio

Constructs a Ratio from numerator and denominator.

Panics

Panics if denom is zero.

fn from_f32(f: f32) -> Option<Ratio>

Creates a Ratio with the value of the given f32. Returns None if the value cannot be converted.

fn from_f64(f: f64) -> Option<Ratio>

Creates a Ratio with the value of the given f64. Returns None if the value cannot be converted.

fn from_integer(i: Integer) -> Ratio

Creates a Ratio from an Integer value.

fn to_f32(&self) -> Option<f32>

Returns the Ratio as an f32 value.

fn to_f64(&self) -> Option<f64>

Returns the Ratio as an f64 value.

fn to_integer(&self) -> Integer

Truncates a Ratio and returns the whole portion as an Integer.

fn is_integer(&self) -> bool

Returns whether the Ratio is an integer; i.e. its denominator is 1.

fn abs(&self) -> Ratio

Returns the absolute value of the Ratio.

fn ceil(&self) -> Ratio

Returns the Ratio rounded towards positive infinity.

fn floor(&self) -> Ratio

Returns the Ratio rounded towards negative infinity.

fn fract(&self) -> Ratio

Returns the fractional portion of a Ratio.

fn round(&self) -> Ratio

Returns the Ratio rounded to the nearest integer. Rounds half-way cases away from zero.

fn trunc(&self) -> Ratio

Returns the Ratio rounded towards zero.

fn recip(&self) -> Ratio

Returns the reciprocal of a Ratio.

Panics

Panics if the numerator is zero.

fn numer(&self) -> &Integer

Returns the Ratio's numerator.

fn denom(&self) -> &Integer

Returns the Ratio's denominator.

fn is_zero(&self) -> bool

Returns whether the Ratio is equal to zero.

fn is_negative(&self) -> bool

Returns whether the Ratio is less than zero.

fn is_positive(&self) -> bool

Returns whether the Ratio is greater than zero.

fn zero() -> Ratio

Returns a Ratio of value zero.

fn one() -> Ratio

Returns a Ratio of value one.