Struct ketos::module::ModuleRegistry

pub struct ModuleRegistry {
    // some fields omitted
}

Loads modules into the running program and caches previously loaded modules

Methods

impl ModuleRegistry

fn new(loader: Box<ModuleLoader>) -> ModuleRegistry

Creates a new ModuleRegistry using the given ModuleLoader to load new modules.

fn get_module(&self, name: Name, scope: &Scope) -> Result<Module, Error>

Returns a loaded module. If the module has not been loaded in this registry; the contained ModuleLoader instance will be used to load it.