Enum ketos::string_fmt::FormatError

pub enum FormatError {
    Error(Box<str>),
    ExtraFields,
    FieldOverflow,
    FieldTypeError {
        expected: &'static str,
        found: &'static str,
    },
    IncompleteDirective,
    IncorrectCloseDelim(char),
    IncorrectFlags,
    InfiniteLoop,
    InsufficientArguments,
    IntegerOverflow,
    InvalidFlags,
    InvalidRadix(u32),
    MisplacedDirective,
    MissingCloseDelim(char),
    MissingBranch,
    ExtraBranch,
    TypeError {
        expected: &'static str,
        found: &'static str,
    },
    UnrecognizedDirective(char),
}

Represents an error in formatting a string.

Variants

Error

Error message

ExtraFields

Extraneous fields to directive

FieldOverflow

Field value exceeded expected range

FieldTypeError

Wrong type to field value

Fields

expected

Type of value expected

found

Type found

IncompleteDirective

End of format string reached before directive completed

IncorrectCloseDelim

Group-closing directive encountered without matching opener

IncorrectFlags

Incorrect flags given to directive

InfiniteLoop

Infinite loop detected in iteration directive ~{...~}

InsufficientArguments

Insufficient arguments to format string

IntegerOverflow

Overflow converting a value to Integer

InvalidFlags

Invalid flags specifier

InvalidRadix

Invalid radix to ~r directive

MisplacedDirective

Directive encountered where it is not expected

MissingCloseDelim

End of format string reached with open grouping directive

MissingBranch

Missing required branch within a conditional directive

ExtraBranch

End of branch directive found where end of conditional was expected

TypeError

Argument value of incorrect type received

Fields

expected

Expected value type

found

Type found

UnrecognizedDirective

Unrecognized directive character

Methods

impl FormatError

fn expected(ty: &'static str, v: &Value) -> FormatError

Convenience function to return a TypeError value when expected type is expected, but some other type of value is found.