Enum ketos::value::Value

pub enum Value {
    Unit,
    Unbound,
    Bool(bool),
    Float(f64),
    Integer(Integer),
    Ratio(Ratio),
    Struct(Rc<Struct>),
    StructDef(Rc<StructDef>),
    Name(Name),
    Keyword(Name),
    Char(char),
    String(String),
    Quasiquote(Box<Value>, u32),
    Comma(Box<Value>, u32),
    CommaAt(Box<Value>, u32),
    Quote(Box<Value>, u32),
    List(RcVec<Value>),
    Function(Function),
    Lambda(Lambda),
    Foreign(Rc<ForeignValue>),
}

Represents a value.

Variants

Unit

()

Unbound

Placeholder for missing optional values; should not be used to represent any real value.

Bool

Boolean -- true or false

Float

Floating point number

Integer

Signed arbitrary precision integer

Ratio

Arbitrary precision ratio

Struct

Struct value

StructDef

Struct definition

Name

Literal name

Keyword

Keyword

Char

Character

String

String

Quasiquote

Quasiquoted value; quote depth MUST NEVER be zero.

Comma

Comma'd value; comma depth MUST NEVER be zero.

CommaAt

Comma-at'd value; quote depth MUST NEVER be zero.

Quote

Quoted value; quote depth MUST NEVER be zero.

List

Series of one or more values. MUST NEVER be of length zero. Use Unit to represent empty lists.

Function

Function implemented in Rust

Lambda

Compiled bytecode function

Foreign

Boxed value of a foreign type

Methods

impl Value

fn new_foreign<T: ForeignValue>(t: T) -> Value

Returns a value of a foreign type.

fn new_foreign_fn<F>(name: Name, f: F) -> Value where F: Any + Fn(&Scope, &mut [Value]) -> Result<Value, Error>

Returns a value containing a foreign function.

fn compare(&self, rhs: &Value) -> Result<Ordering, ExecError>

Compares two values; returns an error if the values cannot be compared.

If a true, Ord-like comparison cannot be made, ExecError::CannotCompare(..) should be returned.

fn is_equal(&self, rhs: &Value) -> Result<bool, ExecError>

Tests two values for equality; returns an error if the values cannot be compared.

fn is_identical(&self, rhs: &Value) -> bool

Returns whether this value is identical to another. The notable difference between this and eq is that float NaN values will compare equal.

fn take(&mut self) -> Value

Replaces the value with Unit and returns the old value.

fn quasiquote(self, n: u32) -> Value

Returns the value, quasi-quoted.

fn comma(self, n: u32) -> Value

Returns the value, comma'd.

fn comma_at(self, n: u32) -> Value

Returns the value, comma-at'd.

fn quote(self, n: u32) -> Value

Returns the value, quoted.

fn type_name(&self) -> &'static str

Returns a string describing the type of the value.